+++ index.html
@@ -7,6 +7,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <script src = "functions.js"></script>
+  <script src="main.js" type="module"></script>
   <title>Кекстаграм</title>
 </head>
